Add pre create wallet state for better dealing with errors and UI
